modify service

+++ src/main/java/com/aliyun/learnjava/service/ScoreService.java
@@ -27,7 +27,7 @@ public class ScoreService {
     }
     
     public Score inserScore(
-        long  studentId,
+        int  studentId,
         int chinese,
         int math,
         int english
